+++ src/sql/warnings.py
@@ -1,2 +1,6 @@
 class JupySQLQuotedNamedParametersWarning(UserWarning):
     pass
+
+
+class JupySQLRollbackPerformed(UserWarning):
+    pass