+++ src/tests/test_column_guesser.py
@@ -1,6 +1,7 @@
 import pytest
 
 from sql.magic import SqlMagic
+from sql import _current
 from IPython.core.interactiveshell import InteractiveShell
 
 ip = InteractiveShell()
@@ -20,6 +21,7 @@ sql_env = SqlEnv("sqlite://")
 @pytest.fixture
 def tbl():
     sqlmagic = SqlMagic(shell=ip)
+    _current._set_sql_magic(sqlmagic)
     ip.register_magics(sqlmagic)
     creator = """
         DROP TABLE IF EXISTS manycoltbl;